creating another file LCAresolver.php for simple lca test case
