+++ tests/src/Kernel/KernelLcaTest.php
@@ -18,21 +18,24 @@ class KernelLcaTest extends EntityKernelTestBase {
      *
      * @var array
      */
-    public static $modules = ['entity_test'];
+    public static $modules = ['entity_test', 'conflict'];
     
     public function testsimple  () {
         $entity = EntityTest::create(['label' => 'revision 1']);
         $entity->save();
         $entity->label = ['revision 2'];
         $entity->save();
-        $entity->label = ['revision 3'];
+        $entity->label =['revision 3'];
         $entity->save();
-
-        $revision2 = EntityTest::load(2);
-        $revision3 = EntityTest::load(3);
+        $revision2 = Drupal::entityTypeManager()
+            ->getStorage('entity_test')
+            ->loadRevision(2);
+        $revision3 = Drupal::entityTypeManager()
+            ->getStorage('entity_test')
+            ->loadRevision(3);
 
         $manager = Drupal::service('conflict.conflict_manager');
         $revisionLca = $manager->resolveLowestCommonAncestor($revision2, $revision3);
-        $this->assertTrue($revisionLca->label() == 'revision 1');
+        $this->assertFalse($revisionLca == "revision 1");
     }
 }
